refactor(server)Renamed Competition to ArenaStates
